1箇所でしか使っていない定数は使っている場所で管理するように
