--- src/consts/align.ts
@@ -1,2 +1,0 @@
-export const RATIO_W = 16
-export const RATIO_H = 9
