+++ src/foundations/align.ts
@@ -1,10 +1,13 @@
-import { RATIO_H, RATIO_W } from "@/consts/align"
 import { OptionFormContents, SizeFormContents } from "@/types/form"
 import { PlaylistBase } from "@/types/playlist"
 import { expansion, range, zip } from "@/util/arrays"
 import { canvas2URL, createCanvas } from "@/util/canvas"
 import fetchPlaylistMaster from "./fetchPlaylistMaster"
 import { parseThumbnailUrl } from "./parseThumbnailUrl"
+
+const RATIO_W = 16
+const RATIO_H = 9
+
 
 export default async function align (playlistBases: PlaylistBase[], size: SizeFormContents, option: OptionFormContents) {
     const { outputWidth, outputHeight, columns, rows } = size
