+++ src/foundations/insertRelatedSongs.ts
@@ -1,5 +1,4 @@
-import { SONG_TYPE_CHECKERS } from "@/consts/playlist"
-import { Song } from "@/types/playlist"
+import { Checker, Song, SongType } from "@/types/playlist"
 import { zipAll } from "@/util/arrays"
 import { getThumbnailUrl } from "@/util/image"
 import matchByChekers from "./matchByChekers"
@@ -22,3 +21,14 @@ export default function insertRelatedSongs (songs: Song[], description: string):
 
     return [...songs, ...relateds].sort((a, b) => ordersMap.get(a)! - ordersMap.get(b)!)
 }
+
+export const SONG_TYPE_CHECKERS = [
+    {
+        type: "youtube",
+        regexp: /https:\/\/www\.youtube\.com\/watch\?v=([-\w]+)/g
+    },
+    {
+        type: "youtube",
+        regexp: /https:\/\/youtu\.be\/([-\w]+)/g
+    }
+] satisfies Checker<SongType>[]
