+++ src/types/playlist.ts
@@ -1,5 +1,5 @@
 import { PLAYLIST_TYPES, SONG_TYPES } from "@/consts/playlist"
-import { Values, ValuesObj } from "./util"
+import { Values } from "./util"
 
 export type PlaylistTypes = Values<typeof PLAYLIST_TYPES>
 
@@ -15,7 +15,7 @@ export type Checker<T> = {
 
 export type FetchPlaylist = (listId: string) => Promise<Playlist>
 
-export type SongType = ValuesObj<typeof SONG_TYPES>
+export type SongType = Values<typeof SONG_TYPES>
 
 export type Playlist = {
     type: PlaylistTypes,
