+++ src/util/image.ts
@@ -1,4 +1,3 @@
-import { SONG_TYPES } from "@/consts/playlist"
 import { SongType } from "@/types/playlist"
 
 const cache = new Map<string, HTMLImageElement>()
@@ -21,7 +20,7 @@ export const getImage = async (url: string) => new Promise<HTMLImageElement>((re
 
 export const getThumbnailUrl = (type: SongType, id: string) => {
     switch (type) {
-        case SONG_TYPES.YOUTUBE: {
+        case "youtube": {
             return `https://img.youtube.com/vi/${id}/default.jpg`
         }
     }
